move common types to lofire crate
