+++ lofire/src/types.rs
@@ -3,12 +3,13 @@
 //! Corresponds to the BARE schema
 
 use serde::{Deserialize, Serialize};
+use std::hash::Hash;
 
 /// 32-byte Blake3 hash digest
 pub type Blake3Digest32 = [u8; 32];
 
 /// Hash digest
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub enum Digest {
     Blake3Digest32(Blake3Digest32),
 }
@@ -17,7 +18,7 @@ pub enum Digest {
 pub type ChaCha20Key = [u8; 32];
 
 /// Symmetric cryptographic key
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub enum SymKey {
     ChaCha20Key(ChaCha20Key),
 }
@@ -55,7 +56,7 @@ pub type Timestamp = u32;
 pub const EPOCH_AS_UNIX_TIMESTAMP: u64 = 1645568520;
 
 /// Relative time (e.g. delay from current time)
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub enum RelTime {
     Seconds(u8),
     Minutes(u8),
@@ -67,6 +68,78 @@ pub enum RelTime {
 // STORAGE OBJECTS
 //
 
-/// Object ID
+/// Object ID:
 /// BLAKE3 hash over the serialized Object with encrypted content
 pub type ObjectId = Digest;
+
+/// Object reference
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
+pub struct ObjectRef {
+    /// Object ID
+    pub id: ObjectId,
+
+    /// Key for decrypting the Object
+    pub key: SymKey,
+}
+
+//
+// COMMON DATA TYPES FOR MESSAGES
+//
+
+/// Peer ID: public key of node
+pub type PeerId = PubKey;
+
+/// IPv4 address
+pub type IPv4 = [u8; 4];
+
+/// IPv6 address
+pub type IPv6 = [u8; 16];
+
+/// IP address
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum IP {
+    IPv4(IPv4),
+    IPv6(IPv6),
+}
+
+/// IP transport protocol
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum IPTransportProtocol {
+    TLS,
+    QUIC,
+}
+
+/// IP transport address
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct IPTransportAddr {
+    pub ip: IP,
+    pub port: u16,
+    pub protocol: IPTransportProtocol,
+}
+
+/// Network address
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum NetAddr {
+    IPTransport(IPTransportAddr),
+}
+
+/// Bloom filter (variable size)
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BloomFilter {
+    /// Number of hash functions
+    pub k: u32,
+
+    /// Filter
+    #[serde(with = "serde_bytes")]
+    pub f: Vec<u8>,
+}
+
+/// Bloom filter (128 B)
+///
+/// (m=1024; k=7; p=0.01; n=107)
+pub type BloomFilter128 = [[u8; 32]; 4];
+
+/// Bloom filter (1 KiB)
+///
+/// (m=8192; k=7; p=0.01; n=855)
+pub type BloomFilter1K = [[u8; 32]; 32];
